chore(cli): test the behaviour of OptionManager on the naming strategy
Those tests will be required before removing GSon from this class. To test the
expected messages we need to add methods for those messages in OutputUtils and
pass that class in each instance of the OptionManager.
With OptionManager relying on OutputUtils to write messages, we can mock the
latest to verify the behaviour of OptionManager.

Relates to https://github.com/OpenAPITools/openapi-style-validator/issues/241