feat(lib): accept specification without path
Prevent a NullPointerException when there is not paths. Since path is not
mandatory there is no need to fail in that case. This commit will handle
that case but validate that there is at least one "path" or one "component"
defined into the specification. And report an (new) OpenAPI error when none
are present.