190 fails when no path (#251)
* Fix 178, should not fail when there are no 'example' for a component

* Update lib/src/main/java/org/openapitools/openapistylevalidator/OpenApiSpecStyleValidator.java

little reformat

* feat(lib): accept specification without path

Prevent a NullPointerException when there is not paths. Since path is not
mandatory there is no need to fail in that case. This commit will handle
that case but validate that there is at least one "path" or one "component"
defined into the specification. And report an (new) OpenAPI error when none
are present.

* chore(lib): enforce the presence of "components" or "paths" in tests

Two tests validate the presence/absence of those sections. We clarify those
expectations by setting them explicitly.

* chore(lib): use Junit's assertDoesNotThrow for better assertions

* chore(cli): use Junit's assertAll to group two assertions on same result

* chore(lib): rename logMissingPathsOrComponents to logMissingPathsAndComponents

The spec is invalid when there is no paths AND no components. But valid when at
least one.

Co-authored-by: Jean-François Côté <jcote@stingray.com>